# artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hetqml-service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/predict": {
      "post": {
        "tags": [
          "predict"
        ],
        "summary": "Predict",
        "operationId": "predict_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/predict/batch": {
      "post": {
        "tags": [
          "predict"
        ],
        "summary": "Predict Batch",
        "operationId": "predict_batch_predict_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchPredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Submit Job",
        "operationId": "submit_job_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/manifest/embedding/{manifest_id}": {
      "get": {
        "tags": [
          "manifest"
        ],
        "summary": "Read Embedding Manifest",
        "operationId": "read_embedding_manifest_manifest_embedding__manifest_id__get",
        "parameters": [
          {
            "name": "manifest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Manifest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingManifest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/manifest/feature-pipeline/{manifest_id}": {
      "get": {
        "tags": [
          "manifest"
        ],
        "summary": "Read Feature Pipeline Manifest",
        "operationId": "read_feature_pipeline_manifest_manifest_feature_pipeline__manifest_id__get",
        "parameters": [
          {
            "name": "manifest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Manifest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeaturePipelineManifest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/manifest/model/{manifest_id}": {
      "get": {
        "tags": [
          "manifest"
        ],
        "summary": "Read Model Manifest",
        "operationId": "read_model_manifest_manifest_model__manifest_id__get",
        "parameters": [
          {
            "name": "manifest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Manifest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelManifest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/evaluations": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "List Evals",
        "operationId": "list_evals_evaluations_get",
        "parameters": [
          {
            "name": "model_manifest_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Manifest Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvaluationRecord"
                  },
                  "title": "Response List Evals Evaluations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/evaluations/{evaluation_id}": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "Read Evaluation",
        "operationId": "read_evaluation_evaluations__evaluation_id__get",
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BatchPredictRequest": {
        "properties": {
          "pairs": {
            "items": {
              "$ref": "#/components/schemas/PredictRequest"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Pairs"
          }
        },
        "type": "object",
        "required": [
          "pairs"
        ],
        "title": "BatchPredictRequest"
      },
      "BatchPredictResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/BatchItemResult"
            },
            "type": "array",
            "title": "Results"
          },
          "summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "results",
          "summary"
        ],
        "title": "BatchPredictResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobRecord": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "submitted_at": {
            "type": "number",
            "title": "Submitted At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "request": {
            "$ref": "#/components/schemas/PredictRequest"
          },
          "response": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PredictResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "quantum_mode": {
            "$ref": "#/components/schemas/QuantumMode"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "tenant_id",
          "status",
          "submitted_at",
          "request",
          "quantum_mode"
        ],
        "title": "JobRecord"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus"
      },
      "JobSubmitRequest": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/PredictRequest"
          },
          "callback_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Callback Url"
          }
        },
        "type": "object",
        "required": [
          "request"
        ],
        "title": "JobSubmitRequest"
      },
...
      "PredictMethod": {
        "type": "string",
        "enum": [
          "classical",
          "quantum_strict",
          "quantum_preferred",
          "auto"
        ],
        "title": "PredictMethod"
      },
      "PredictRequest": {
        "properties": {
          "drug_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Drug Id",
            "description": "Exact DrugBank ID. No fuzzy matching."
          },
          "disease_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Disease Id",
            "description": "Exact DOID identifier."
          },
          "method": {
            "$ref": "#/components/schemas/PredictMethod",
            "default": "auto"
          }
        },
        "type": "object",
        "required": [
          "drug_id",
          "disease_id"
        ],
        "title": "PredictRequest"
      },
      "PredictResponse": {
        "properties": {
          "drug_id": {
            "type": "string",
            "title": "Drug Id"
          },
          "disease_id": {
            "type": "string",
            "title": "Disease Id"
          },
          "probability": {
            "type": "number",
            "title": "Probability"
          },
          "method_requested": {
            "$ref": "#/components/schemas/PredictMethod"
          },
          "method_used": {
            "$ref": "#/components/schemas/PredictMethod"
          },
          "quantum_mode_used": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuantumMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          },
          "manifest_chain": {
            "$ref": "#/components/schemas/ManifestChain"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          }
        },
        "type": "object",
        "required": [
          "drug_id",
          "disease_id",
          "probability",
          "method_requested",
          "method_used",
          "manifest_chain",
          "request_id",
          "tenant_id"
        ],
        "title": "PredictResponse"
      },
      "QuantumMode": {
        "type": "string",
        "enum": [
          "simulator",
          "gpu_simulator",
          "ibm_heron",
          "unavailable"
        ],
        "title": "QuantumMode"
      },
...
    }
  }
}